fix: jubjubi by user position

+++ Jubging/src/main/java/Capstone_team1/Jubging/service/JubjubiService.java
@@ -55,12 +55,13 @@ public class JubjubiService {
         String[] coord = userPosition.split(" ");
         String[] northWestLatLng = coord[0].split(",");
         String[] southEastLatLng = coord[1].split(",");
-        float maxLat = Float.parseFloat(northWestLatLng[0]);
-        float minLng = Float.parseFloat(northWestLatLng[1]);
-        float minLat = Float.parseFloat(southEastLatLng[0]);
-        float maxLng = Float.parseFloat(southEastLatLng[1]);
+        double maxLat = Double.parseDouble(northWestLatLng[0]);
+        double minLng = Double.parseDouble(northWestLatLng[1]);
+        double minLat = Double.parseDouble(southEastLatLng[0]);
+        double maxLng = Double.parseDouble(southEastLatLng[1]);
 
         List<JubjubiResponseDto> jubjubiNearByList = jubjubiRepository.findByUserPosition(minLat, minLng, maxLat, maxLng);
+
         return jubjubiNearByList;
     }
